Moved storage engine test util to main storage engine module
